chore: move fn out of constants file
